Vector: Fix allocation bug in ensure_capacity
Previously a vector of larger size would seg fault because of the
condition on whether the vector should grow or not. Rather than checking
if the size is equal to the capacity, it instead should check if the
capacity is greater than or equal to the size, and if so, return as
normal.