List: Add a remove method
Removes all nodes in the list that store the value passed. Uses a helper
function m_delete_node that will handle various circumstances for node
deletion. This includes: when the node to be deleted is the head node,
when the node to be deleted is the tail node, when the list is of m_size
1 and the node is to be deleted.